+++ packages/core/src/js/index.ts
@@ -48,6 +48,31 @@ export {
   addEventProcessor,
   lastEventId,
   consoleSandbox,
+  instrumentOpenAiClient,
+  instrumentAnthropicAiClient,
+  instrumentGoogleGenAIClient,
+  createLangChainCallbackHandler,
+  instrumentLangGraph,
+  instrumentStateGraphCompile,
+} from '@sentry/core';
+
+export type {
+  OpenAiClient,
+  OpenAiOptions,
+  InstrumentedMethod,
+  AnthropicAiClient,
+  AnthropicAiOptions,
+  AnthropicAiInstrumentedMethod,
+  AnthropicAiResponse,
+  GoogleGenAIClient,
+  GoogleGenAIChat,
+  GoogleGenAIOptions,
+  GoogleGenAIIstrumentedMethod,
+  LangChainOptions,
+  LangChainIntegration,
+  LangGraphOptions,
+  LangGraphIntegration,
+  CompiledGraph,
 } from '@sentry/core';
 
 export {